W2 implemented, working depth buffer.

+++ source/Camera.h
@@ -82,11 +82,11 @@ namespace dae
 			float cameraSpeed{ 0.2f };
 
 			const uint8_t* pKeyboardState = SDL_GetKeyboardState(nullptr);
-			origin += (pKeyboardState[SDL_SCANCODE_W] | pKeyboardState[SDL_SCANCODE_UP]) * forward * deltaTime * movementSpeed;
-			origin += (pKeyboardState[SDL_SCANCODE_S] | pKeyboardState[SDL_SCANCODE_DOWN]) * -forward * deltaTime * movementSpeed;
+			origin += static_cast<float>((pKeyboardState[SDL_SCANCODE_W] | pKeyboardState[SDL_SCANCODE_UP])) * forward * deltaTime * movementSpeed;
+			origin += static_cast<float>((pKeyboardState[SDL_SCANCODE_S] | pKeyboardState[SDL_SCANCODE_DOWN])) * -forward * deltaTime * movementSpeed;
 
-			origin += (pKeyboardState[SDL_SCANCODE_A] | pKeyboardState[SDL_SCANCODE_LEFT]) * -right * deltaTime * movementSpeed;
-			origin += (pKeyboardState[SDL_SCANCODE_D] | pKeyboardState[SDL_SCANCODE_RIGHT]) * right * deltaTime * movementSpeed;
+			origin += static_cast<float>((pKeyboardState[SDL_SCANCODE_A] | pKeyboardState[SDL_SCANCODE_LEFT])) * -right * deltaTime * movementSpeed;
+			origin += static_cast<float>((pKeyboardState[SDL_SCANCODE_D] | pKeyboardState[SDL_SCANCODE_RIGHT])) * right * deltaTime * movementSpeed;
 
 			int mouseX{}, mouseY{};
 			const uint32_t mouseState = SDL_GetRelativeMouseState(&mouseX, &mouseY);
@@ -95,13 +95,13 @@ namespace dae
 			{
 			case SDL_BUTTON_X2:
 			{
-				origin += up * mouseY * deltaTime * movementSpeed;
+				origin += up * static_cast<float>(mouseY) * deltaTime * movementSpeed;
 
 				break;
 			}
 			case SDL_BUTTON_LMASK:
 			{
-				origin += forward * mouseY * deltaTime * movementSpeed;
+				origin += forward * static_cast<float>(mouseY) * deltaTime * movementSpeed;
 
 				totalYaw += mouseX * cameraSpeed;
 				Matrix rotation = Matrix::CreateRotation(totalPitch * TO_RADIANS, 0, 0);
